fix:ip location

+++ app/api-node/src/migration/1591705594611-geoip.ts
@@ -27,12 +27,12 @@ export class geoip1591705594611 implements MigrationInterface {
                 },
                 {
                     "name": "latitude",
-                    "type": "integer",
+                    "type": "numeric",
                     "isNullable": false
                 },
                 {
                     "name": "longitude",
-                    "type": "integer",
+                    "type": "numeric",
                     "isNullable": false
                 },
                 {
